requirements.txt
pytest
click
python-dotenv